adds new ingredients_reference table and updates existing tables

+++ src/main/resources/db.migration/V20240416_2045__create_user_table.sql
@@ -1,5 +1,5 @@
 CREATE TABLE `users` (
-    `user_id` int NOT NULL AUTO_INCREMENT PRIMARY KEY,
+    `user_id` BIGINT NOT NULL AUTO_INCREMENT PRIMARY KEY,
     `email` varchar(50),
     `first_name` varchar(20),
     `second_name` varchar(20)
